+++ previous-versions/common/codatcommon.go
@@ -164,9 +164,9 @@ func New(opts ...SDKOption) *CodatCommon {
 		sdkConfiguration: sdkConfiguration{
 			Language:          "go",
 			OpenAPIDocVersion: "3.0.0",
-			SDKVersion:        "0.27.0",
-			GenVersion:        "2.159.2",
-			UserAgent:         "speakeasy-sdk/go 0.27.0 2.159.2 3.0.0 github.com/codatio/client-sdk-go/previous-versions/common",
+			SDKVersion:        "0.28.0",
+			GenVersion:        "2.169.0",
+			UserAgent:         "speakeasy-sdk/go 0.28.0 2.169.0 3.0.0 github.com/codatio/client-sdk-go/previous-versions/common",
 		},
 	}
 	for _, opt := range opts {